Experimental fix for bugs #36888.
When the flash video object is forced to the backaground, it is
cloned, removed and the clone reinserted with fake type
(x-fake/x-flash-stoppped) to ensure that the flash plugin stops
playing on the background.

When the flash video object is brought to the foreground it is cloned,
removed and the clone reinserted with valid type
(application/x-shockwave-flash), so a flash plugin can handle it.

When the Linterna Mágica video object is sent to the background, its
playback is stopped.

#future_release_0.0.13