A work-around for bugs #39402 https://savannah.nongnu.org/bugs/?39402
A message is displayed to the user that the clips is not playable,
because of encrypted signatures (or missing) in YouTube clips. A link
to the bug report is provided. This is not a proper fix.

According to the bug report, the encryption/decryption algorithm
changes every few days. As of this commit it seems that the bug is not
present when tested via a proxy.